+++ src/components/card/card.css
@@ -21,7 +21,7 @@
       width: 50px;
       color: #feda4a;
       fill: #feda4a; }
-    .card .top-card .imperial.favorited {
+    .card .top-card .favorited {
       background-color: #feda4a; }
   .card .card-data {
     background-color: #f8f9f4;
