+++ src/components/card/card.scss
@@ -25,7 +25,7 @@
       color: #feda4a;
       fill: #feda4a;
     }
-    .imperial.favorited {
+    .favorited {
       background-color: #feda4a;
     }
   }
